+++ lib/preface_stdlib/either.mli
@@ -43,6 +43,12 @@ module Applicative (T : Preface_specs.Types.T0) :
   Preface_specs.Traversable.API_OVER_APPLICATIVE
     with type 'a t = (T.t, 'a) Bifunctor.t
 
+(** {3 Selective} *)
+
+module Selective (T : Preface_specs.Types.T0) :
+  Preface_specs.SELECTIVE
+    with type 'a t = (T.t, 'a) Bifunctor.t
+
 (** {3 Monad}
 
     [Either.t] implements {!module-type:Preface_specs.MONAD} and introduces an