+++ lib/preface_stdlib/option.ml
@@ -67,6 +67,11 @@ module Monad_traversable (M : Preface_specs.MONAD) =
 module Monad =
   Preface_make.Traversable.Join_with_monad (Monad_internal) (Monad_traversable)
 
+module Selective =
+  Preface_make.Selective.Over_applicative_via_select
+    (Applicative)
+    (Preface_make.Selective.Select_from_monad (Monad))
+
 module Monad_plus =
   Preface_make.Monad_plus.Over_monad_and_alternative (Monad) (Alternative)