+++ lib/preface_stdlib/try.mli
@@ -26,6 +26,10 @@ module Alt : Preface_specs.ALT with type 'a t = 'a t
 module Applicative :
   Preface_specs.Traversable.API_OVER_APPLICATIVE with type 'a t = 'a t
 
+(** {2 Selective} *)
+
+module Selective : Preface_specs.SELECTIVE with type 'a t = 'a t
+
 (** {2 Monad}
 
     [Try.t] implements {!module-type:Preface_specs.MONAD} and introduces an